files_uploader/requirements.txt
bandit==1.7.5
celery==5.3.4
django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-split-settings==1.2.0
djangorestframework==3.14.0
flake8==6.1.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==23.7.10
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-html==0.4.3
flake8-isort==6.0.0
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
gunicorn==21.2.0
pillow==10.1.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
wemake-python-styleguide==0.18.0
